Import platform relations
See discussion in https://github.com/hiddewie/OpenRailwayMap-vector/discussions/278#discussioncomment-12971807 about Münster (https://www.openstreetmap.org/node/3374898279) not being qualified as a large station.

One of the problems is that `public_transport=platform` are only imported if they are tagged as nodes or ways, not if they are relations.