Ignore signal feature type for popup
Fixes #533

The feature type (none, tram or line) for signals was used in the popup. This is not correct for signals, as the code expects `point` or `line` as feature type. All signals are points (the default).

Only features without icon matching were impacted. Signals matching on possible signal state, or speed signals matching on speeds, were not impacted.