Detect unknown signals
Currently icons are shown for unknown signal types, e.g. an unknown main signal.

However, signals that do not have a signal type are not shown on the map at all. This pull request renders the unknown signal type for those signals.

In Poland there are 616 such signals, of the 44126 total signals. For example https://www.openstreetmap.org/node/278819312.